feat(BudgetValidator): 예산 검증 로직을 BudgetValidator로 분리
BudgetValidator 클래스를 새로 생성하여 예산 검증 로직을 분리하였습니다.
기존의 LottoValidator에서 예산 검증 메서드를 제거하고 Budget 클래스에서 해당 클래스를 사용하도록 수정하였습니다.
이로 인해 코드의 가독성과 유지보수성이 향상되었습니다.
